+++ tests/unit/test_sonata/test_split_population.py
@@ -258,110 +258,6 @@ def test_simple_split_subcircuit(tmp_path):
         assert list(group["source_node_id"]) == [1]
         assert list(group["target_node_id"]) == [0]
 
-
-def test__gather_layout_from_networks():
-    res = split_population._gather_layout_from_networks({"nodes": [], "edges": []})
-    assert res == ({}, {})
-
-    nodes, edges = split_population._gather_layout_from_networks(
-        {
-            "nodes": [
-                {
-                    "nodes_file": "a/b/a.h5",
-                    "populations": {"a": {"type": "biophysical"}},
-                },
-                {
-                    "nodes_file": "a/b/bc.h5",
-                    "populations": {"b": {"type": "biophysical"}, "c": {"type": "biophysical"}},
-                },
-                {
-                    "nodes_file": "a/b/a.h5",
-                    "populations": {"A": {"type": "biophysical"}},
-                },
-            ],
-            "edges": [
-                {
-                    "edges_file": "a/b/a.h5",
-                    "populations": {"a_a": {"type": "biophysical"}},
-                },
-                {
-                    "edges_file": "a/b/bc.h5",
-                    "populations": {"b_c": {"type": "biophysical"}, "c_b": {"type": "biophysical"}},
-                },
-                {
-                    "edges_file": "a/a/bc.h5",
-                    "populations": {"a_c": {"type": "biophysical"}, "a_b": {"type": "biophysical"}},
-                },
-                {
-                    "edges_file": "a/b/a.h5",
-                    "populations": {"A_a": {"type": "biophysical"}},
-                },
-            ],
-        }
-    )
-    assert nodes == {
-        "A": "A/a.h5",
-        "a": "a/a.h5",
-        "b": "b/bc.h5",
-        "c": "b/bc.h5",
-    }
-    assert edges == {
-        "A_a": "A_a/a.h5",
-        "a_a": "a_a/a.h5",
-        "a_b": "a/bc.h5",
-        "a_c": "a/bc.h5",
-        "b_c": "b/bc.h5",
-        "c_b": "b/bc.h5",
-    }
-
-    nodes, edges = split_population._gather_layout_from_networks(
-        {
-            "nodes": [
-                {
-                    "nodes_file": "a/b/bc.h5",
-                    "populations": {"b": {"type": "biophysical"}, "c": {"type": "biophysical"}},
-                },
-                {
-                    "nodes_file": "a/b/bc.h5",
-                    "populations": {"B": {"type": "biophysical"}, "C": {"type": "biophysical"}},
-                },
-            ],
-            "edges": [],
-        }
-    )
-    assert nodes == {"B": "b/bc.h5", "C": "b/bc.h5", "b": "b/bc.h5", "c": "b/bc.h5"}
-
-    nodes, edges = split_population._gather_layout_from_networks(
-        {
-            "nodes": [
-                {
-                    "nodes_file": "a/b/a.h5",
-                    "populations": {"a": {"type": "biophysical"}},
-                },
-                {
-                    "nodes_file": "a/b/bc.h5",
-                    "populations": {"b": {"type": "biophysical"}},
-                },
-                {
-                    "nodes_file": "a/b/bc.h5",
-                    "populations": {"c": {"type": "biophysical"}},
-                },
-                {
-                    "nodes_file": "a/b/a.h5",
-                    "populations": {"A": {"type": "biophysical"}},
-                },
-            ],
-            "edges": [],
-        }
-    )
-    assert nodes == {
-        "A": "A/a.h5",
-        "a": "a/a.h5",
-        "b": "b/bc.h5",
-        "c": "c/bc.h5",
-    }
-
-
 def test__update_node_sets():
     ret = split_population._update_node_sets(node_sets={}, id_mapping={})
     assert ret == {}
